+++ src/ts/aias.ts
@@ -1,9 +1,10 @@
 
 import { Is } from '@lcluber/chjs';
 import { Logger } from '@lcluber/mouettejs';
-import { HTTPHeaderFields } from './httpheaderfields';
+import { HTTPHeaders } from './httpheaders';
 
 export type HTTPRequestMethods = 'GET'|'HEAD'|'POST'|'PUT'|'DELETE'|'CONNECT'|'OPTIONS'|'TRACE'|'PATCH';
+export type dataTypes = string|Document|Blob|BufferSource|FormData|ArrayBufferView|ArrayBuffer|FormData|URLSearchParams|ReadableStream|null;
 
 export class HTTP {
 
@@ -11,8 +12,9 @@ export class HTTP {
   // static method: HTTPRequestMethods;
   static async: boolean = true;
   static noCache: boolean = false;
-  static headers: HTTPHeaderFields = {
-    'Content-Type': 'application/json'//'application/x-www-form-urlencoded'
+  static base64: boolean = false;
+  static headers: HTTPHeaders = {
+    //'Content-Type': 'application/x-www-form-urlencoded'//'application/json'
   };
 
   public static get( url: string ): Promise<string> {
@@ -23,11 +25,11 @@ export class HTTP {
     return this.call('HEAD', url);
   }
 
-  public static post( url: string, data: string|Document|Blob|ArrayBufferView|ArrayBuffer|FormData|URLSearchParams|ReadableStream ): Promise<string> {
+  public static post( url: string, data: dataTypes ): Promise<string> {
     return this.call('POST', url, data);
   }
 
-  public static put( url: string, data: string|Document|Blob|ArrayBufferView|ArrayBuffer|FormData|URLSearchParams|ReadableStream ): Promise<string> {
+  public static put( url: string, data: dataTypes ): Promise<string> {
     return this.call('PUT', url, data);
   }
 
@@ -47,11 +49,11 @@ export class HTTP {
     return this.call('TRACE', url);
   }
 
-  public static patch( url: string, data: string|Document|Blob|ArrayBufferView|ArrayBuffer|FormData|URLSearchParams|ReadableStream ): Promise<string> {
+  public static patch( url: string, data: dataTypes ): Promise<string> {
     return this.call('PATCH', url, data);
   }
 
-  public static setHeaders(headers: HTTPHeaderFields): void {
+  public static setHeader(headers: HTTPHeaders): void {
     for(const property in headers){
       if (headers.hasOwnProperty(property)) {
         this.headers[property] = headers[property];
@@ -59,49 +61,46 @@ export class HTTP {
     }
   }
 
-  private static call( method: HTTPRequestMethods, url: string, data?: string|Document|Blob|ArrayBufferView|ArrayBuffer|FormData|URLSearchParams|ReadableStream ): Promise<string> {
+  private static call( method: HTTPRequestMethods, url: string, data?: dataTypes): Promise<string> {
     return new Promise((resolve: Function, reject: Function) => {
 
+      let msg = ['Aias xhr ', ' ('+ method +':' + url + ')'];
       let http = new XMLHttpRequest();
-
-      if(this.noCache) {
-        url += '?cache=' + (new Date()).getTime();
-      }
+      
+      url += this.noCache ? '?cache=' + (new Date()).getTime() : '';
 
       http.open(method, url, this.async);
-
-      for (let property in this.headers) {
-        if (this.headers.hasOwnProperty(property)) {
-          http.setRequestHeader(property, this.headers[property]);
-        }
-      }
+      
+      this.setRequestHeaders(http);
 
       http.onreadystatechange = () => {
         if(http.readyState == 4) {
           if(http.status == 200) {
-            Logger.info('xhr done successfully ('+url+')');
+            Logger.info(msg[0] + 'successful' + msg[1]);
             resolve(http.responseText);
           } else {
-            Logger.error('xhr failed ('+url+')');
+            Logger.error(msg[0] + 'failed' + msg[1]);
             reject(http.status);
           }
         }
       };
-      Logger.info('xhr processing starting ('+url+')');
-      if (data == undefined){
-        http.send();
-        return;
-      }
-      //
-      // let contentType = 'application/json';
-      // if (Check.isString(data)) {
-      //   contentType = 'application/x-www-form-urlencoded';
+
       if (Is.object(data)) {
         data = JSON.stringify(data);
       }
-      // http.setRequestHeader('Content-Type', contentType);
-      http.send(data);
+
+      http.send(data || null);
+      Logger.info(msg[0] + 'sent' + msg[1]);
+
     });
   }
-
+  
+  private static setRequestHeaders(http: XMLHttpRequest): void {
+    for (let property in this.headers) {
+      if (this.headers.hasOwnProperty(property)) {
+        http.setRequestHeader(property, this.headers[property]);
+      }
+    }
+  }
+  
 }